+++ .betterer.results
@@ -6506,6 +6506,20 @@ exports[`better eslint`] = {
       [173, 23, 47, "Do not use any type assertions.", "3309878203"],
       [195, 43, 45, "Do not use any type assertions.", "15355460"]
     ],
+    "public/app/features/plugins/sql/components/visual-query-builder/AwesomeQueryBuilder.tsx:760854115": [
+      [175, 24, 63, "Do not use any type assertions.", "2252455532"],
+      [198, 13, 47, "Do not use any type assertions.", "2763495851"],
+      [253, 30, 3, "Unexpected any. Specify a different type.", "193409811"]
+    ],
+    "public/app/features/plugins/sql/datasource/SqlDatasource.ts:178321817": [
+      [124, 33, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [159, 56, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [177, 19, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [182, 57, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [184, 18, 135, "Do not use any type assertions.", "3270957200"],
+      [194, 28, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [225, 33, 3, "Unexpected any. Specify a different type.", "193409811"]
+    ],
     "public/app/features/plugins/tests/datasource_srv.test.ts:2399414445": [
       [10, 19, 3, "Unexpected any. Specify a different type.", "193409811"],
       [44, 48, 21, "Do not use any type assertions.", "932413114"],
@@ -10512,20 +10526,6 @@ exports[`better eslint`] = {
       [78, 19, 3, "Unexpected any. Specify a different type.", "193409811"],
       [113, 33, 3, "Unexpected any. Specify a different type.", "193409811"],
       [117, 19, 3, "Unexpected any. Specify a different type.", "193409811"]
-    ],
-    "public/app/plugins/datasource/sql/components/visual-query-builder/AwesomeQueryBuilder.tsx:1840701731": [
-      [175, 24, 63, "Do not use any type assertions.", "2252455532"],
-      [198, 13, 47, "Do not use any type assertions.", "2763495851"],
-      [253, 30, 3, "Unexpected any. Specify a different type.", "193409811"]
-    ],
-    "public/app/plugins/datasource/sql/datasource/SqlDatasource.ts:3607048882": [
-      [121, 33, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [156, 56, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [174, 19, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [179, 57, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [181, 18, 135, "Do not use any type assertions.", "3270957200"],
-      [191, 28, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [222, 33, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
     "public/app/plugins/datasource/tempo/QueryEditor/NativeSearch.test.tsx:3510665018": [
       [32, 18, 160, "Do not use any type assertions.", "2324324215"],
