+++ public/app/features/plugins/sql/components/ConfirmModal.tsx
