+++ public/app/features/plugins/sql/components/query-editor-raw/QueryValidator.tsx
