+++ public/app/features/plugins/sql/components/visual-query-builder/AwesomeQueryBuilder.tsx
