+++ public/app/features/plugins/sql/components/visual-query-builder/GroupByRow.tsx
