+++ public/app/features/plugins/sql/constants.ts
