+++ public/app/features/plugins/sql/defaults.ts
