+++ public/app/features/plugins/sql/types.ts
