+++ public/app/features/plugins/sql/utils/formatSQL.ts
