+++ public/app/features/plugins/sql/utils/sql.utils.ts
