+++ public/app/features/plugins/sql/utils/useSqlChange.ts
